refactor(theme): refactored currency swithcer
