+++ packages/theme/components/CurrencySelector/CurrenciesModal.vue
@@ -1,0 +1,126 @@
+<template>
+  <SfBottomModal
+    :is-open="isModalOpen"
+    title="Choose Currency"
+    @click:close="closeModal"
+  >
+    <SfList
+      v-if="availableCurrencies.length > 1"
+    >
+      <SfListItem
+        v-for="currency in availableCurrencies"
+        :key="currency"
+      >
+        <a
+          href="/"
+          :class="selectedCurrency === currency ? 'container__currency--selected-label' : ''"
+          @click.prevent="handleChanges({
+            callback: () => changeCurrency({id: currency}),
+            redirect: false,
+            refresh: true
+          })"
+        >
+          <SfCharacteristic class="currency">
+            <template #title>
+              <span>{{ currency }}</span>
+            </template>
+          </SfCharacteristic>
+        </a>
+      </SfListItem>
+    </SfList>
+  </SfBottomModal>
+</template>
+<script>
+import {
+  defineComponent, computed, onMounted,
+} from '@nuxtjs/composition-api';
+import {
+  SfList,
+  SfBottomModal,
+  SfCharacteristic,
+} from '@storefront-ui/vue';
+import {
+  useCurrency,
+} from '@vue-storefront/magento';
+import { useHandleChanges } from '~/helpers/magentoConfig/handleChanges';
+
+export default defineComponent({
+  name: 'CurrenciesModal',
+  components: {
+    SfList,
+    SfBottomModal,
+    SfCharacteristic,
+  },
+  props: {
+    isModalOpen: Boolean,
+    selectedCurrency: String,
+  },
+  emits: ['closeModal'],
+  setup() {
+    const {
+      currencies,
+      change: changeCurrency,
+      load: loadCurrencies,
+    } = useCurrency('header-currency');
+
+    const { handleChanges } = useHandleChanges();
+
+    const availableCurrencies = computed(() => currencies.value?.available_currency_codes || []);
+
+    onMounted(() => {
+      if (currencies.value && currencies.value?.available_currency_codes) return;
+      loadCurrencies();
+    });
+
+    return {
+      changeCurrency,
+      handleChanges,
+      availableCurrencies,
+    };
+  },
+  methods: {
+    closeModal() {
+      this.$emit('closeModal');
+    },
+  },
+});
+</script>
+<style lang="scss" scoped>
+.sf-bottom-modal {
+  z-index: 2;
+  left: 0;
+  @include for-desktop {
+    --bottom-modal-height: 100vh;
+  }
+}
+.sf-bottom-modal::v-deep .sf-bottom-modal__close {
+  position: var(--circle-icon-position, absolute);
+  top: var(--spacer-xs);
+  right: var(--spacer-xs);
+}
+
+.sf-list {
+  .language {
+    padding: var(--spacer-sm);
+
+    &__flag {
+      margin-right: var(--spacer-sm);
+    }
+  }
+
+  @include for-desktop {
+    display: flex;
+  }
+}
+
+.container__currency {
+  &--selected-label {
+    font-weight: bold;
+  }
+
+  &:hover,
+  &--selected {
+    opacity: 1;
+  }
+}
+</style>
